# examples/reference-app/internal/notes/api/notes-server.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// NotesCreateNoteRequest defines model for NotesCreateNoteRequest.
type NotesCreateNoteRequest struct {
	// Text The note's placeholder content.
	Text string `json:"text"`
}

// NotesError The structured {code, params} error envelope every speed API returns instead of localized text: code is a stable machine-readable identifier naming exactly one failure, and params carries the structured per-code details when the failure has any. The server never renders a message -- a client maps each code to its own locale's text.
type NotesError struct {
	// Code Example: notes.text_required
	Code   *string                 `json:"code,omitempty"`
	Params *map[string]interface{} `json:"params,omitempty"`
}

// NotesListNotesResponse defines model for NotesListNotesResponse.
type NotesListNotesResponse struct {
	Notes *[]NotesNote `json:"notes,omitempty"`
}

// NotesNote defines model for NotesNote.
type NotesNote struct {
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// ID Application-generated UUID.
	ID   *string `json:"id,omitempty"`
	Text *string `json:"text,omitempty"`
}

// NotesCreateNoteJSONRequestBody defines body for NotesCreateNote for application/json ContentType.
type NotesCreateNoteJSONRequestBody = NotesCreateNoteRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// NotesListNotes List every note belonging to the caller's tenant.
	// (GET /api/v1/notes)
	NotesListNotes(w http.ResponseWriter, r *http.Request)
	// NotesCreateNote Create a note for the caller's tenant.
	// (POST /api/v1/notes)
	NotesCreateNote(w http.ResponseWriter, r *http.Request)
	// NotesDeleteNote Mark a note as deleted for the caller's tenant.
	// (DELETE /api/v1/notes/{noteId})
	NotesDeleteNote(w http.ResponseWriter, r *http.Request, noteID string)
	// NotesRestoreNote Restore a previously deleted note for the caller's tenant.
	// (POST /api/v1/notes/{noteId}/restore)
	NotesRestoreNote(w http.ResponseWriter, r *http.Request, noteID string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// NotesListNotes operation middleware
func (siw *ServerInterfaceWrapper) NotesListNotes(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.NotesListNotes(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// NotesCreateNote operation middleware
func (siw *ServerInterfaceWrapper) NotesCreateNote(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.NotesCreateNote(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// NotesDeleteNote operation middleware
func (siw *ServerInterfaceWrapper) NotesDeleteNote(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "noteId" -------------
	var noteID string

	err = runtime.BindStyledParameterWithOptions("simple", "noteId", r.PathValue("noteId"), &noteID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "noteId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.NotesDeleteNote(w, r, noteID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// NotesRestoreNote operation middleware
func (siw *ServerInterfaceWrapper) NotesRestoreNote(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "noteId" -------------
	var noteID string

	err = runtime.BindStyledParameterWithOptions("simple", "noteId", r.PathValue("noteId"), &noteID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "noteId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.NotesRestoreNote(w, r, noteID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/notes", wrapper.NotesListNotes)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/notes", wrapper.NotesCreateNote)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/api/v1/notes/{noteId}", wrapper.NotesDeleteNote)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/notes/{noteId}/restore", wrapper.NotesRestoreNote)

	return m
}