# examples/reference-app/web/app-openapi.yaml
openapi: 3.0.3
info:
  title: reference-app
  description: "The reference-app application OpenAPI document: the redocly join of the application own three fragments (notes, cases and smilesim), regenerated by `task api:gen:app` and committed. Each fragment stays authoritative for its own operations; this merged document is the app-owned SDK generation input and the linted whole-surface view."
  version: 0.0.0
...
paths:
  /api/v1/notes:
    servers: []
    post:
      operationId: notes_createNote
      tags:
        - notes
      summary: Create a note for the caller's tenant.
      description: >-
        The tenant is always the one tenancy.Middleware already resolved from
        the request; there is no tenant_id field anywhere on this request -- the
        tenant comes from the resolved request context, never from anything the
        caller sends.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesCreateNoteRequest'
      responses:
        '201':
          description: The note was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesNote'
        '400':
          description: >-
            The request body was missing, malformed, had empty text, or text
            exceeding the maxLength declared below.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesError'
    get:
      operationId: notes_listNotes
      tags:
        - notes
      summary: List every note belonging to the caller's tenant.
      responses:
        '200':
          description: The caller's tenant's notes, oldest registration order first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesListNotesResponse'
  /api/v1/notes/{noteId}:
    servers: []
    delete:
      operationId: notes_deleteNote
      tags:
        - notes
      summary: Mark a note as deleted for the caller's tenant.
      description: >-
        The tenant is always the one tenancy.Middleware already resolved from
        the request, and the note id names a note belonging to that tenant alone
        -- never a tenant_id request field: the tenant comes from the resolved
        request context, never from anything the caller sends. The delete is a
        mark-delete (dbkit.SoftDeletable): the row stays in place with
        deleted_at/deleted_by set, hidden from the list operation, and
        restorable through notes_restoreNote until a retention sweep physically
        reaps it. Deleting an already-deleted note answers 404 exactly like an
        unknown or another tenant's id -- the refusal is uniform, so a caller
        cannot learn which case it hit.
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
          description: The note's application-generated UUID.
      responses:
        '204':
          description: The note was marked deleted.
        '401':
          description: >-
            The request could not be attributed to a user
            (notes.subject_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesError'
        '404':
          description: >-
            No live note of the caller's tenant carries that id -- the note does
            not exist, belongs to another tenant, or is already deleted
            (notes.note_not_found). The answer is uniform across the three
            cases, exactly like the 404 the repository's other operations answer
            for a cross-tenant id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesError'
  /api/v1/notes/{noteId}/restore:
    servers: []
    post:
      operationId: notes_restoreNote
      tags:
        - notes
      summary: Restore a previously deleted note for the caller's tenant.
      description: >-
        The inverse of notes_deleteNote: clears the mark-delete's
        deleted_at/deleted_by so the note is visible to the list operation
        again, with its text and creator untouched. Restoring a note that is not
        currently deleted -- because it was never deleted, is unknown to this
        tenant, or belongs to another tenant -- answers 404 uniformly
        (notes.note_not_found), so a caller cannot learn which case it hit.
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
          description: The note's application-generated UUID.
      responses:
        '204':
          description: The note was restored.
        '401':
          description: >-
            The request could not be attributed to a user
            (notes.subject_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesError'
        '404':
          description: >-
            No currently-deleted note of the caller's tenant carries that id --
            the note is live, does not exist, or belongs to another tenant
            (notes.note_not_found). The answer is uniform across the three
            cases.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesError'
...
components:
  schemas:
    NotesCreateNoteRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 4000
          description: The note's placeholder content.
    NotesNote:
      type: object
      properties:
        id:
          type: string
          description: Application-generated UUID.
        text:
          type: string
        created_at:
          type: string
          format: date-time
    NotesListNotesResponse:
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: '#/components/schemas/NotesNote'
    NotesError:
      type: object
      description: >-
        The structured {code, params} error envelope every speed API returns
        instead of localized text: code is a stable machine-readable identifier
        naming exactly one failure, and params carries the structured per-code
        details when the failure has any. The server never renders a message --
        a client maps each code to its own locale's text.
      properties:
        code:
          type: string
          example: notes.text_required
        params:
          type: object
          additionalProperties: true
...